Renamed  ImagePlaceholder -> SkeletonCard and moved to shared components
