+++ apps/verification-portal/src/lib/shared/components/SkeletonCard.svelte
