+++ apps/verification-portal/src/lib/widgets/NFTClaim/ClaimData.svelte
@@ -4,7 +4,7 @@
   import CardHeader from '$lib/shared/typography/CardHeader.svelte';
   import Property from '$lib/shared/typography/Property.svelte';
   import NftImage from '$lib/components/NFTImage.svelte';
-  import ImagePlaceholder from '$lib/components/ImagePlaceholder.svelte';
+  import SkeletonCard from '$lib/shared/components/SkeletonCard.svelte';
   import SocialFollowMenu from '$lib/entities/SocialFollow/SocialFollowMenu.svelte';
 
   const { claimResponse, claimStatus } = getNFTClaimContext();
@@ -65,7 +65,7 @@
       />
     {:else}
       <div class="hidden md:block w-48 h-48 rounded-lg overflow-hidden">
-        <ImagePlaceholder className="w-48 h-48 block "></ImagePlaceholder>
+        <SkeletonCard className="w-48 h-48 block "></SkeletonCard>
       </div>
     {/if}
     <div>
